refactor: MainPage 폴더로 분리
